21 Aralık // 16.45

+++ src/main/java/com/example/newProject/api/controllers/HouseController.java
@@ -25,9 +25,11 @@ public class HouseController {
     }
 
     @GetMapping("/getAllHousesOfOneCustomer/{customerId}")
-    public List<House> getAllHousesOfOneCustomer(@PathVariable int customerId){
+    public List<House> getAllHousesOfOneCustomer(@PathVariable Long customerId){
         return houseService.getAllHousesOfOneCustomer(customerId);
     }
+
+
